+++ CodeIndigo/Indigo/Text.h
@@ -9,7 +9,7 @@ public:
 
 	// Create new text based off of the text. -1 is bottom, 1 is top. Text is placed with it's top-left corner at (text_x, text_y).
 	// If time_to_last is negative, text won't disappear.
-	Text(const float text_x=0, const float text_y=0, char * display="", float * color = nullptr, void * font = nullptr, const int time_to_last = -1, void(*finished)(void)=nullptr);
+	Text(const float text_x=0, const float text_y=0, char * display="", float * color = nullptr, void * font = nullptr, const float time_to_last = -1, void(*finished)(void)=nullptr);
 	// Create new Text based off of another Text object
 	Text(const Text& copy);
 	// Destroy the text
@@ -25,8 +25,8 @@ private:
 	char * text;
 	// The typeset to use
 	void * typeset;
-	// The frames until the text should go away
-	int finished;
+	// The time until the text should go away
+	float finished;
 	// The position of the text in actual pixels
 	float x;
 	float y;
